requirements.txt
# 核心依赖
tushare>=1.4.0
pandas>=2.0.0
numpy>=1.24.0

# 数据库
duckdb>=0.10.0
sqlalchemy>=2.0.0

# UI框架（Streamlit路线A）
streamlit>=1.30.0
plotly>=5.18.0
streamlit-aggrid>=0.3.4

# 工具库
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
retry>=0.9.2

# 日期处理
python-dateutil>=2.8.2

# 日志
loguru>=0.7.0

# 开发工具（可选）
ipython>=8.12.0
jupyter>=1.0.0
